split the interface
